test: reorganized to address amman client and rest client
