+++ amman-tests/tests/rest-client/queries.ts
@@ -1,0 +1,8 @@
+import spok from 'spok'
+import test from 'tape'
+import { killAmman, launchAmman, relayClient } from '../utils/launch'
+
+// test('rest-client: given amman is running with the relay enabled', async (t) => {
+//   const client = relayClient()
+//   const state = await launchAmman()
+// })
