+++ amman-tests/tests/rest-client/tasks/restart-validator.ts
@@ -1,0 +1,5 @@
+import test from 'tape'
+
+test('rest:restart-validator', (t) => {
+  t.pass('coming up')
+})
